Put configuration of each element into own file
