+++ application/lib/model/ConfigState.dart
@@ -1,0 +1,9 @@
+import 'package:application/view/screens/configurationScreen.dart';
+import 'package:flutter/material.dart';
+
+abstract class ConfigState extends State<ConfigurationScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return null;
+  }
+}
